refactor(modules): relocate coding-credentials into src/modules/coding/credentials
Behavior-preserving move: coding-credentials source (6 files) and tests
(8 files) relocated into src/modules/coding/credentials/, with import
paths updated in the moved files and all external importers
(module.ts, coding-secrets-facade.ts, debug/settings routes, and their
tests). The DB schema file src/db/coding-credentials-schema.ts stays in
place. No logic changes.
