+++ src/debug/settings/coding-credentials-fields-meta.ts
@@ -3,7 +3,13 @@
 // Use of this software is governed by the Business Source License 1.1.
 // See LICENSE in the project root for details.
 
-import { AGENTS, AUTH_METHODS, FORGE_KINDS, PROVIDERS, type CodingNamespace } from '../../coding-credentials/types.js'
+import {
+  AGENTS,
+  AUTH_METHODS,
+  FORGE_KINDS,
+  PROVIDERS,
+  type CodingNamespace,
+} from '../../modules/coding/credentials/types.js'
 
 export type FieldMeta = {
   key: string
