+++ src/modules/coding/credentials/mcp-selections.ts
