+++ tests/debug/settings/coding-credentials-routes.test.ts
@@ -7,11 +7,11 @@ import { beforeEach, describe, expect, test } from 'bun:test'
 
 import { z } from 'zod'
 
-import { setMcpCatalog } from '../../../src/coding-credentials/mcp-catalog.js'
-import { parseMcpSelections } from '../../../src/coding-credentials/mcp-selections.js'
-import { getCodingCredentials } from '../../../src/coding-credentials/store.js'
 import { maskSensitiveValue } from '../../../src/config.js'
 import { handleCodingCredentialsRoutes } from '../../../src/debug/settings/coding-credentials-routes.js'
+import { setMcpCatalog } from '../../../src/modules/coding/credentials/mcp-catalog.js'
+import { parseMcpSelections } from '../../../src/modules/coding/credentials/mcp-selections.js'
+import { getCodingCredentials } from '../../../src/modules/coding/credentials/store.js'
 import { resolveSettingsPrincipal } from '../../../src/settings/principal.js'
 import { addUser } from '../../../src/users.js'
 import { mockLogger, seedTestPlatformInstance, setupTestDb } from '../../utils/test-helpers.js'
