+++ tests/modules/coding/credentials/resolve-agent-secrets.test.ts
@@ -7,13 +7,18 @@ import { afterEach, beforeEach, expect, test } from 'bun:test'
 
 import { eq } from 'drizzle-orm'
 
-import { addAuthorizedGroup } from '../../src/authorized-groups.js'
+import { addAuthorizedGroup } from '../../../../src/authorized-groups.js'
 import {
   getConfigContextIdFromStorageContextId,
   toScopedContextId,
   toScopedThreadContextId,
-} from '../../src/chat/scoped-context.js'
-import { adminCodingGuardrailsContextId, setCodingGuardrails } from '../../src/coding-credentials/guardrails.js'
+} from '../../../../src/chat/scoped-context.js'
+import { getDrizzleDb } from '../../../../src/db/drizzle.js'
+import { authorizedGroups } from '../../../../src/db/schema.js'
+import {
+  adminCodingGuardrailsContextId,
+  setCodingGuardrails,
+} from '../../../../src/modules/coding/credentials/guardrails.js'
 import {
   resolveAgent,
   resolveAgentSecrets,
@@ -21,11 +26,9 @@ import {
   resolveForgeToken,
   resolveModel,
   resolveProviderHost,
-} from '../../src/coding-credentials/resolve-agent-secrets.js'
-import { updateCodingCredentials } from '../../src/coding-credentials/store.js'
-import { getDrizzleDb } from '../../src/db/drizzle.js'
-import { authorizedGroups } from '../../src/db/schema.js'
-import { mockLogger, setupTestDb } from '../utils/test-helpers.js'
+} from '../../../../src/modules/coding/credentials/resolve-agent-secrets.js'
+import { updateCodingCredentials } from '../../../../src/modules/coding/credentials/store.js'
+import { mockLogger, setupTestDb } from '../../../utils/test-helpers.js'
 
 const STORAGE_CTX = 'pi:telegram:ctx:user-9'
 // STORAGE_CTX is not parseable (non-standard format) → pi === undefined → legacy path unchanged
@@ -186,7 +189,7 @@ test('resolveForge returns null for a partial self-hosted vault (instance_url pr
 })
 
 // MCP resolution (resolveMcpServers/resolveMcpTokens) is covered by
-// tests/coding-credentials/resolve-mcp-servers.test.ts.
+// tests/modules/coding/credentials/resolve-mcp-servers.test.ts.
 
 test('resolveProviderHost returns null when no credentials stored', () => {
   expect(resolveProviderHost(STORAGE_CTX, 'user-9')).toBeNull()
